modified:   source/fixes/AladdinInNasirasRevengeWidescreenFix.cpp 	modified:   source/fixes/AlcatrazPrisonEscapeFOVFix.cpp 	modified:   source/fixes/AmericanMcGeesAliceFOVFix.cpp 	modified:   source/fixes/Americas10MostWantedFOVFix.cpp 	modified:   source/fixes/AnneMcCaffreysFreedomFirstResistanceFOVFix.cpp 	modified:   source/fixes/AntzExtremeRacingWidescreenFix.cpp 	modified:   source/fixes/ArthursQuestBattlefortheKingdomFOVFix.cpp 	modified:   source/fixes/AtlantisTheLostEmpireTrialbyFireFOVFix.cpp 	modified:   source/fixes/BarbieHorseAdventuresMysteryRideWidescreenFix.cpp 	modified:   source/fixes/BarbieHorseAdventuresRidingCampWidescreenFix.cpp 	modified:   source/fixes/BarbieSparklingIceShowWidescreenFix.cpp 	modified:   source/fixes/BarbieandtheMagicofPegasusWidescreenFix.cpp 	modified:   source/fixes/BattleIsleTheAndosiaWarFOVFix.cpp 	modified:   source/fixes/BeetleCrazyCupFOVFix.cpp 	modified:   source/fixes/BeyondNormandyAssignmentBerlinWidescreenFix.cpp 	modified:   source/fixes/BigMuthaTruckers2WidescreenFix.cpp 	modified:   source/fixes/BigMuthaTruckersFOVFix.cpp 	modified:   source/fixes/BillyBladeTempleofTimeFOVFix.cpp 	modified:   source/fixes/CTUMarineSharpshooterFOVFix.cpp 	modified:   source/fixes/Cabelas4x4OffroadAdventure3FOVFix.cpp 	modified:   source/fixes/ChameleonWidescreenFix.cpp 	modified:   source/fixes/CircusEmpireFOVFix.cpp 	modified:   source/fixes/CrisisTeamAmbulanceDriverFOVFix.cpp 	new file:   source/fixes/CurseTheEyeOfIsisFOVFix.cpp 	modified:   source/fixes/DieHardNakatomiPlazaFOVFix.cpp 	modified:   source/fixes/DrBrainActionReactionFOVFix.cpp 	modified:   source/fixes/DromeRacersFOVFix.cpp 	modified:   source/fixes/EliteForcesNavySEALsFOVFix.cpp 	modified:   source/fixes/EliteForcesNavySEALsWeaponsofMassDestructionFOVFix.cpp 	modified:   source/fixes/EliteForcesWW2IwoJimaFOVFix.cpp 	modified:   source/fixes/EliteForcesWW2NormandyFOVFix.cpp 	new file:   source/fixes/F1WorldGrandPrix2000WidescreenFix.cpp 	modified:   source/fixes/FroggerBeyondFOVFix.cpp 	modified:   source/fixes/HarleyDavidsonRacetotheRallyFOVFix.cpp 	modified:   source/fixes/HugoBukkazoomWidescreenFix.cpp 	modified:   source/fixes/KISSPsychoCircusTheNightmareChildFOVFix.cpp 	modified:   source/fixes/LarryRaglands4x4ChallengeFOVFix.cpp 	modified:   source/fixes/LegendsofMightandMagicFOVFix.cpp 	modified:   source/fixes/LegoRacers2WidescreenFix.cpp 	modified:   source/fixes/MadeManConfessionsOfTheFamilyBloodWidescreenFix.cpp 	modified:   source/fixes/MagicTheGatheringBattlegroundsFOVFix.cpp 	modified:   source/fixes/MiamiViceFOVChanger.cpp 	modified:   source/fixes/MightandMagic9FOVFix.cpp 	modified:   source/fixes/MistmareFOVFix.cpp 	modified:   source/fixes/NinaAgentChroniclesFOVFix.cpp 	modified:   source/fixes/OperationSandstormFOVChanger.cpp 	modified:   source/fixes/ProjectIGI1WidescreenFix.cpp 	modified:   source/fixes/RevolutionFOVFix.cpp 	modified:   source/fixes/RuneGoldCutscenesFOVFix.cpp 	modified:   source/fixes/SASAntiTerrorForceFOVChanger.cpp 	modified:   source/fixes/SOEOperationAvalancheFOVChanger.cpp 	modified:   source/fixes/SanityAikensArtifactFOVFix.cpp 	modified:   source/fixes/SearchandRescueCoastalHeroesFOVFix.cpp 	modified:   source/fixes/SearchandRescueVietnamMedEvacFOVFix.cpp 	modified:   source/fixes/SecretServiceInHarmsWayFOVFix.cpp 	modified:   source/fixes/ShadowForceRazorUnitFOVFix.cpp 	modified:   source/fixes/Spiderman2TheGameFOVFix.cpp 	modified:   source/fixes/SplatMagazineRenegadePaintballFOVFix.cpp 	modified:   source/fixes/Starsky&HutchFOVFix.cpp 	modified:   source/fixes/Swat3FOVChanger.cpp 	modified:   source/fixes/TNNOutdoorsProHunter2FOVFix.cpp 	modified:   source/fixes/TheHistoryChannelCivilWarANationDividedFOVFix.cpp 	modified:   source/fixes/TheMummyWidescreenFix.cpp 	modified:   source/fixes/TheSumofAllFearsFOVFix.cpp 	modified:   source/fixes/USMostWantedNowhereToHideFOVFix.cpp 	modified:   source/fixes/VietnamBlackOpsFOVFix.cpp 	modified:   source/fixes/WesternOutlawWantedDeadorAliveFOVFix.cpp 	modified:   source/fixes/WingsOfWarFOVFix.cpp

+++ source/fixes/BeyondNormandyAssignmentBerlinWidescreenFix.cpp
@@ -96,7 +96,7 @@ void HandleFOVInput(double &newCustomFOVMultiplier)
 		replace(input.begin(), input.end(), ',', '.');
 
 		// Parses the string to a double
-		newCustomFOVMultiplier = stof(input);
+		newCustomFOVMultiplier = stod(input);
 
 		if (cin.fail())
 		{
