+++ source/fixes/DromeRacersFOVFix.cpp
@@ -82,7 +82,7 @@ double HandleFOVInput()
         replace(input.begin(), input.end(), ',', '.');
 
         // Parses the string to a double
-        newCustomFOVInDegrees = stof(input);
+        newCustomFOVInDegrees = stod(input);
 
         if (cin.fail())
         {
