+++ source/fixes/HugoBukkazoomWidescreenFix.cpp
@@ -92,7 +92,7 @@ void HandleFOVInput(double &customFOV)
         replace(input.begin(), input.end(), ',', '.');
 
         // Parses the string to a double
-        customFOV = stof(input);
+        customFOV = stod(input);
 
         if (cin.fail())
         {
