+++ source/fixes/SASAntiTerrorForceFOVChanger.cpp
@@ -69,7 +69,7 @@ double HandleFOVInput()
         replace(input.begin(), input.end(), ',', '.');
 
         // Parses the string to a double
-        customFOV = stof(input);
+        customFOV = stod(input);
 
         if (cin.fail())
         {
